backend/requirements.txt
blinker==1.9.0
certifi==2025.11.12
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pymongo==4.15.4
python-dotenv==1.2.1
Werkzeug==3.1.3
